[REFACTOR] #18: 관리자 관련 코드 AdminService로 이동
